refactor(server side rendering): cleanup & restucture of server side rendering - moved out most of the code to seperate folder - split out various helpers - added test files - updated tests with cleanup
